AutoMapper, CheckResult, Attribute validation

+++ AllAboutGames.Data/Models/Forum/ForumLike.cs
@@ -7,17 +7,17 @@ namespace AllAboutGames.Data.Models.Forum
     public class ForumLike
     {
         [Key]
-        public int ForumLikeID { get; set; }
+        public long ForumLikeID { get; set; }
 
         [Required]
         [ForeignKey(nameof(User))]
-        public int UserID { get; set; }
+        public long UserID { get; set; }
 
         public ApplicationUser User { get; set; }
 
         [Required]
         [ForeignKey(nameof(ForumPost))]
-        public int ForumPostID { get; set; }
+        public long ForumPostID { get; set; }
 
         public ForumPost ForumPost { get; set; }
     }
